feat(profile/contact forms) added custom hooks for  profile and contact forms
